reduced bvh memory size

+++ Luminary/lib/processing.c
@@ -17,7 +17,7 @@ static float linearRGB_to_SRGB(const float value) {
 void frame_buffer_to_8bit_image(Camera camera, raytrace_instance* instance, RGB8* image) {
   RGBF* error_table = (RGBF*) malloc(sizeof(RGBF) * (instance->width + 2));
 
-  assert((int) error_table, "Failed to allocate memory!", 1);
+  assert((unsigned long long) error_table, "Failed to allocate memory!", 1);
 
   memset(error_table, 0, sizeof(RGBF) * (instance->width + 2));
 
@@ -66,7 +66,7 @@ void frame_buffer_to_8bit_image(Camera camera, raytrace_instance* instance, RGB8
 void frame_buffer_to_16bit_image(Camera camera, raytrace_instance* instance, RGB16* image) {
   RGBF* error_table = (RGBF*) malloc(sizeof(RGBF) * (instance->width + 2));
 
-  assert((int) error_table, "Failed to allocate memory!", 1);
+  assert((unsigned long long) error_table, "Failed to allocate memory!", 1);
 
   memset(error_table, 0, sizeof(RGBF) * (instance->width + 2));
